Refactor authentication tests: consolidate into dedicated test files
- Moved registration and login tests from handlers_test.go to auth_test.go.
- Removed handlers_test.go as it was redundant after the refactor.
- Added health check tests to health_test.go.
- Created helpers_test.go for shared setup functions and test utilities.
- Added password reset and token verification tests to password_test.go and token_test.go respectively.
- Ensured all tests are properly structured and utilize the new setup functions.
