+++ api/auth-service/src/health_test.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestHealthCheck(t *testing.T) {
+	router := setupTestRouter()
+
+	req, _ := http.NewRequest("GET", "/health", nil)
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, req)
+
+	assert.Equal(t, http.StatusOK, w.Code)
+
+	var response map[string]interface{}
+	require.NoError(t, json.Unmarshal(w.Body.Bytes(), &response))
+
+	assert.Equal(t, true, response["success"])
+	data := response["data"].(map[string]interface{})
+	assert.Equal(t, "auth-service", data["service"])
+	assert.Equal(t, "healthy", data["status"])
+}
